requirements.txt
argcomplete==1.8.2
beautifulsoup4==4.5.3
certifi==2019.3.9
chardet==3.0.4
cycler==0.10.0
cymem==2.0.2
cytoolz==0.9.0.1
dataclasses==0.6
defusedxml==0.5.0
dill==0.2.8.2
Django==2.1.7
django-allauth==0.39.1
django-crispy-forms==1.7.2
djangorestframework==3.9.2
djongo==1.2.31
docx2txt==0.6
idna==2.8
kiwisolver==1.0.1
lxml==4.2.5
matplotlib==3.0.2
msgpack==0.5.6
msgpack-numpy==0.4.3.2
murmurhash==1.0.1
nltk==3.4
numpy==1.15.4
oauthlib==3.0.1
pandas==0.23.4
pdfminer.six==20181108
Pillow==5.3.0
plac==0.9.6
preshed==2.0.1
pycryptodome==3.7.0
pymongo==3.7.2
pyparsing==2.3.0
PyPDF2==1.26.0
pyreadline==2.1
python-dateutil==2.7.5
python-pptx==0.6.5
python3-openid==3.1.0
pytz==2018.9
regex==2018.1.10
requests==2.21.0
requests-oauthlib==1.2.0
scikit-learn==0.20.1
scipy==1.1.0
singledispatch==3.4.0.3
six==1.10.0
sortedcontainers==2.0.5
spacy==2.0.18
SpeechRecognition==3.6.3
sqlparse==0.2.4
thinc==6.12.1
toolz==0.9.0
tqdm==4.28.1
ujson==1.35
urllib3==1.24.1
wrapt==1.10.11
xlrd==1.0.0
XlsxWriter==1.1.2
